api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dateutil>=2.8.2
drf-spectacular>=0.27.0
django-filter>=23.5
